Moved SpatialOrientation in to sofast dir
