+++ example/fixed_pattern_deflectometry/physical_target_dot_calibration.py
@@ -10,7 +10,7 @@ from opencsp.app.sofast.lib.CalibrateSofastFixedDots import CalibrateSofastFixed
 from opencsp.common.lib.camera.Camera import Camera
 from opencsp.common.lib.deflectometry.CalibrationCameraPosition import \
     CalibrationCameraPosition
-from opencsp.common.lib.deflectometry.SpatialOrientation import SpatialOrientation
+from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
 from opencsp.common.lib.geometry.Vxy import Vxy
 from opencsp.common.lib.geometry.Vxyz import Vxyz
 from opencsp.common.lib.opencsp_path.opencsp_root_path import opencsp_code_dir
