+++ opencsp/app/sofast/lib/ProcessSofastFixed.py
@@ -18,7 +18,7 @@ import opencsp.common.lib.deflectometry.image_processing as ip
 import opencsp.common.lib.deflectometry.process_optics_geometry as pr
 from opencsp.common.lib.deflectometry.SlopeSolver import SlopeSolver
 from opencsp.common.lib.deflectometry.SlopeSolverData import SlopeSolverData
-from opencsp.common.lib.deflectometry.SpatialOrientation import SpatialOrientation
+from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
 from opencsp.common.lib.geometry.RegionXY import RegionXY
 from opencsp.common.lib.geometry.Uxyz import Uxyz
