+++ opencsp/app/sofast/lib/SpatialOrientation.py
